+++ sktree/tree/unsupervised/_unsup_splitter.pxd
@@ -1,5 +1,5 @@
 from ..._lib.sklearn.tree._splitter cimport BaseSplitter, SplitRecord
-from ..._lib.sklearn.utils._typedefs cimport float64_t, float32_t, intp_t, uint32_t
+from ..._lib.sklearn.utils._typedefs cimport float32_t, float64_t, intp_t, uint32_t
 from ._unsup_criterion cimport UnsupervisedCriterion
